# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediBook Multi-Hospital Doctor Appointment System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hospitals:
    get:
      operationId: getHospitals
      tags: [hospitals]
      summary: List all approved hospitals
      parameters:
        - name: approved
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of hospitals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hospital"
    post:
      operationId: createHospital
      tags: [hospitals]
      summary: Register a new hospital
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHospitalBody"
      responses:
        "201":
          description: Hospital registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /hospitals/{id}:
    get:
      operationId: getHospital
      tags: [hospitals]
      summary: Get hospital by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hospital details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateHospital
      tags: [hospitals]
      summary: Update hospital (approve/details)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHospitalBody"
      responses:
        "200":
          description: Hospital updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"
  /doctors:
    get:
      operationId: getDoctors
      tags: [doctors]
      summary: List doctors with optional filters
      parameters:
        - name: departmentId
          in: query
          required: false
          schema:
            type: integer
        - name: hospitalId
          in: query
          required: false
          schema:
            type: integer
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [rating, reviews, availability, score]
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: minRating
          in: query
          required: false
          schema:
            type: number
        - name: maxFee
          in: query
          required: false
          schema:
            type: number
        - name: availableToday
          in: query
          required: false
          schema:
            type: boolean
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of doctors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorListResponse"
    post:
      operationId: createDoctor
      tags: [doctors]
      summary: Create a doctor profile (hospital admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDoctorBody"
      responses:
        "201":
          description: Doctor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get doctor profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateDoctor
      tags: [doctors]
      summary: Update doctor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDoctorBody"
      responses:
        "200":
          description: Doctor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
  /doctors/{id}/reviews:
    get:
      operationId: getDoctorReviews
      tags: [doctors]
      summary: Get reviews for a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
  /doctors/{id}/availability:
    get:
      operationId: getDoctorAvailability
      tags: [availability]
      summary: Get availability slots for a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Availability"
    post:
      operationId: setDoctorAvailability
      tags: [availability]
      summary: Set availability slots for a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAvailabilityBody"
      responses:
        "200":
          description: Availability set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Hospital:
      type: object
      required: [id, name, location, approved]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        approved:
          type: boolean
        phone:
          type: [string, 'null']
        email:
          type: [string, 'null']
        description:
          type: [string, 'null']
        imageUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    CreateHospitalBody:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
        location:
          type: string
        phone:
          type: string
        email:
          type: string
        description:
          type: string
    UpdateHospitalBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        approved:
          type: boolean
        phone:
          type: string
        email:
          type: string
        description:
          type: string
    Doctor:
      type: object
      required: [id, userId, hospitalId, departmentId, experience, consultationFee, averageRating, totalReviews]
      properties:
        id:
          type: integer
        userId:
          type: string
        hospitalId:
          type: integer
        departmentId:
          type: integer
        experience:
          type: integer
        consultationFee:
          type: number
        averageRating:
          type: number
        totalReviews:
          type: integer
        bio:
          type: [string, 'null']
        specialization:
          type: [string, 'null']
        qualification:
          type: [string, 'null']
        isTopRated:
          type: boolean
        nextAvailableSlot:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        email:
          type: [string, 'null']
        hospitalName:
          type: [string, 'null']
        departmentName:
          type: [string, 'null']
    DoctorDetail:
      type: object
      required: [id, userId, hospitalId, departmentId, experience, consultationFee, averageRating, totalReviews]
      properties:
        id:
          type: integer
        userId:
          type: string
        hospitalId:
          type: integer
        departmentId:
          type: integer
        experience:
          type: integer
        consultationFee:
          type: number
        averageRating:
          type: number
        totalReviews:
          type: integer
        bio:
          type: [string, 'null']
        specialization:
          type: [string, 'null']
        qualification:
          type: [string, 'null']
        isTopRated:
          type: boolean
        nextAvailableSlot:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        email:
          type: [string, 'null']
        hospitalName:
          type: [string, 'null']
        departmentName:
          type: [string, 'null']
        recentReviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    DoctorListResponse:
      type: object
      required: [doctors, total, page, limit]
      properties:
        doctors:
          type: array
          items:
            $ref: "#/components/schemas/Doctor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CreateDoctorBody:
      type: object
      required: [userId, hospitalId, departmentId, experience, consultationFee]
      properties:
        userId:
          type: string
        hospitalId:
          type: integer
        departmentId:
          type: integer
        experience:
          type: integer
        consultationFee:
          type: number
        bio:
          type: string
        specialization:
          type: string
        qualification:
          type: string
    UpdateDoctorBody:
      type: object
      properties:
        experience:
          type: integer
        consultationFee:
          type: number
        bio:
          type: string
        specialization:
          type: string
        qualification:
          type: string
    Availability:
      type: object
      required: [id, doctorId, date, timeSlots]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        date:
          type: string
          format: date
        timeSlots:
          type: array
          items:
            type: string
    SetAvailabilityBody:
      type: object
      required: [date, timeSlots]
      properties:
        date:
          type: string
          format: date
        timeSlots:
          type: array
          items:
            type: string
...
    Review:
      type: object
      required: [id, patientId, doctorId, rating, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: string
        doctorId:
          type: integer
        appointmentId:
          type: [integer, 'null']
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        patientFirstName:
          type: [string, 'null']
        patientLastName:
          type: [string, 'null']
        patientProfileImageUrl:
          type: [string, 'null']
    ReviewListResponse:
      type: object
      required: [reviews, total, page, limit]
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...