# lib/api-zod/src/generated/types/doctorDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediBook Multi-Hospital Doctor Appointment System API
 * OpenAPI spec version: 0.1.0
 */
import type { Review } from "./review";

export interface DoctorDetail {
  id: number;
  userId: string;
  hospitalId: number;
  departmentId: number;
  experience: number;
  consultationFee: number;
  averageRating: number;
  totalReviews: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  specialization?: string | null;
  /** @nullable */
  qualification?: string | null;
  isTopRated?: boolean;
  /** @nullable */
  nextAvailableSlot?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  hospitalName?: string | null;
  /** @nullable */
  departmentName?: string | null;
  recentReviews?: Review[];
}